[Silver III] Title: 주유소, Time: 372 ms, Memory: 36464 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -8,20 +8,20 @@ public class Main {
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         int N = Integer.parseInt(br.readLine());
-        
-        int[] distance = new int[N-1];
+
+        long[] distance = new long[N-1];
         StringTokenizer st = new StringTokenizer(br.readLine());
         for (int i = 0; i < N-1; i++) {
-            distance[i] = Integer.parseInt(st.nextToken());
+            distance[i] = Long.parseLong(st.nextToken());
         }
-        int[] cost = new int[N];
+        long[] cost = new long[N];
         st = new StringTokenizer(br.readLine());
         for (int i = 0; i < N; i++) {
-            cost[i] = Integer.parseInt(st.nextToken());
+            cost[i] = Long.parseLong(st.nextToken());
         }
 
-        int sum = 0;
-        int minCost = cost[0];
+        long sum = 0;
+        long minCost = cost[0];
 
         for (int i = 0; i < N-1; i++) {
             if (cost[i] < minCost) {
